Move tests to new package
